Adds the post-download processing functions.

+++ src/datavault_api_client/post_download_processing.py
@@ -32,26 +32,18 @@ sequential order. The newly concatenated files are then passed to the integrity 
 functions and, if any file fails the integrity test, its partitions are added to the list
 of files and partitions whose download is to be repeated.
 """
-
+import copy
+import itertools
 import pathlib
 import shutil
-from typing import Any, List, Tuple, Union
-import itertools
+from typing import Any, List, Optional
+
 from datavault_api_client.data_integrity import get_list_of_failed_downloads
 from datavault_api_client.data_structures import (
     ConcurrentDownloadManifest,
     DownloadDetails,
     PartitionDownloadDetails,
 )
-from datavault_api_client.pre_download_processing import filter_files_to_split
-
-
-# def post_download_processing(
-#     downloaded_files: List[DownloadDetails],
-#     synchronous: bool = False,
-# ) -> List[DownloadDetails]:
-#     if synchronous:
-#         return get_list_of_failed_downloads(downloaded_files)
 
 
 ##########################################################################################
@@ -60,30 +52,56 @@ from datavault_api_client.pre_download_processing import filter_files_to_split
 def get_non_partitioned_files(
     whole_files_reference_data: List[DownloadDetails],
 ) -> List[DownloadDetails]:
+    """Returns a list of non-partitioned files.
+
+    Whether a file is partitioned or not is assessed by checking the is_partitioned flag
+    in the DownloadDetails named-tuple.
+
+    Parameters
+    ----------
+    whole_files_reference_data: List[DownloadDetails]
+        A list of DownloadDetails named tuples containing the information of each downloaded
+        file.
+
+    Returns
+    -------
+    List[DownloadDetails]
+        A list of DownloadDetails named-tuples of all those files where the is_partitioned
+        flag is False.
+    """
     return [file for file in whole_files_reference_data if file.is_partitioned is False]
 
 
 def get_partitioned_files(
     whole_files_reference_data: List[DownloadDetails],
 ) -> List[DownloadDetails]:
-    return list(
-        set(whole_files_reference_data).difference(
-            set(get_non_partitioned_files(whole_files_reference_data))
-        )
-    )
+    """Returns a list of partitioned files.
+
+    Parameters
+    ----------
+    whole_files_reference_data: List[DownloadDetails]
+        A list of DownloadDetails named tuples containing the information of each downloaded
+        file.
+
+    Returns
+    -------
+    List[DownloadDetails]
+        A list of DownloadDetails named-tuples of all those files where the is_partitioned
+        flag is True.
+    """
+    return [file for file in whole_files_reference_data if file.is_partitioned is True]
 
 
 def get_partitions_download_details(
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]],
-    file_name: str = None
+    partitions_to_download: List[PartitionDownloadDetails],
+    file_name: Optional[str] = None,
 ) -> List[PartitionDownloadDetails]:
     """Filters from the download manifest the PartitionDownloadDetails named-tuples.
 
     Parameters
     ----------
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]]
-        A list of DownloadDetails and PartitionDownloadDetails named-tuples, constituting
-        the  download manifest of all the whole files and file partitions to download.
+    partitions_to_download: List[PartitionDownloadDetails]
+        A list of PartitionDownloadDetails named-tuples.
     file_name: str
         The name of a file. By default is set to None.
 
@@ -97,33 +115,30 @@ def get_partitions_download_details(
         file.
     """
     if not file_name:
-        return [
-            file for file in concurrent_download_manifest
-            if type(file) is PartitionDownloadDetails
-        ]
+        return partitions_to_download
     return [
-        file for file in concurrent_download_manifest
-        if type(file) is PartitionDownloadDetails and file.parent_file_name == file_name
+        file for file in partitions_to_download
+        if file.parent_file_name == file_name
     ]
 
 
 def get_downloaded_partitions(path_to_folder: pathlib.Path) -> List[Any]:
     """Retrieves the full paths of the partitions file in a folder.
 
-        Parameters
-        ----------
-        path_to_folder: pathlib.Path
-            A pathlib.Path indicating the full path to the directory where we want to check
-            for partition files.
+    Parameters
+    ----------
+    path_to_folder: pathlib.Path
+        A pathlib.Path indicating the full path to the directory where we want to check
+        for partition files.
 
-        Returns
-        -------
-        List[Any]
-            If in the directory that is passed as an input are found partition files, the
-            function will return a list of pathlib.Path objects each containing the full
-            path to an individual partition file. If no partition file is found in the
-            directory, the function will return an empty list.
-        """
+    Returns
+    -------
+    List[Any]
+        If in the directory that is passed as an input are found partition files, the
+        function will return a list of pathlib.Path objects each containing the full
+        path to an individual partition file. If no partition file is found in the
+        directory, the function will return an empty list.
+    """
     downloaded_partitions = list(path_to_folder.glob("*.txt"))
     if len(downloaded_partitions) != 0:
         downloaded_partitions.sort(key=lambda x: int(x.stem.split("_")[3]))
@@ -132,7 +147,7 @@ def get_downloaded_partitions(path_to_folder: pathlib.Path) -> List[Any]:
 
 def get_file_specific_missing_partitions(
     file_specific_download_details: DownloadDetails,
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]],
+    partitions_to_download: List[PartitionDownloadDetails],
 ) -> List[Any]:
     """Compares the downloaded against the expected partitions and returns the missing partitions.
 
@@ -140,8 +155,8 @@ def get_file_specific_missing_partitions(
     ----------
     file_specific_download_details: DownloadDetails
         A DownloadDetails named-tuple containing file-specific download information.
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]]
-        A list of DownloadDetails and PartitionDownloadDetails named-tuples.
+    partitions_to_download: List[PartitionDownloadDetails]
+        A list of PartitionDownloadDetails named-tuples.
 
     Returns
     -------
@@ -152,7 +167,7 @@ def get_file_specific_missing_partitions(
         list.
     """
     expected_partitions = get_partitions_download_details(
-        concurrent_download_manifest,
+        partitions_to_download,
         file_name=file_specific_download_details.file_name,
     )
     return [
@@ -165,7 +180,7 @@ def get_file_specific_missing_partitions(
 
 def get_all_missing_partitions(
     whole_files_reference_data: List[DownloadDetails],
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]],
+    partitions_to_download: List[PartitionDownloadDetails],
 ) -> List[PartitionDownloadDetails]:
     """Returns all the missing partitions from a download session.
 
@@ -174,8 +189,8 @@ def get_all_missing_partitions(
     whole_files_reference_data: List[DownloadDetails]
         A list of DownloadDetails named-tuples each containing file-specific download
         information.
-    concurrent_download_manifest: List[Union[DownloadDetails, PartitionDownloadDetails]]
-        A list of DownloadDetails and PartitionDownloadDetails named-tuples.
+    partitions_to_download: List[PartitionDownloadDetails]
+        A list of PartitionDownloadDetails named-tuples.
 
     Returns
     -------
@@ -187,9 +202,9 @@ def get_all_missing_partitions(
         the function will return a tuple of empty lists.
     """
     missing_partitions = [
-        get_file_specific_missing_partitions(file, concurrent_download_manifest)
+        get_file_specific_missing_partitions(file, partitions_to_download)
         for file in get_partitioned_files(whole_files_reference_data)
-        if len(get_file_specific_missing_partitions(file, concurrent_download_manifest)) > 0
+        if len(get_file_specific_missing_partitions(file, partitions_to_download)) > 0
     ]
     return list(itertools.chain.from_iterable(missing_partitions))
 
@@ -223,7 +238,7 @@ def get_files_with_missing_partitions(
 
 def get_files_ready_for_concatenation(
     whole_files_reference_data: List[DownloadDetails],
-    files_with_missing_partitions: List[DownloadDetails]
+    files_with_missing_partitions: List[DownloadDetails],
 ) -> List[DownloadDetails]:
     """Returns a list of files that are not missing any partition and are ready for concatenation.
 
@@ -248,7 +263,7 @@ def get_files_ready_for_concatenation(
     return list(
         set(get_partitioned_files(whole_files_reference_data)).difference(
             files_with_missing_partitions,
-        )
+        ),
     )
 
 
@@ -267,11 +282,12 @@ def concatenate_partitions(path_to_output_file: pathlib.Path) -> str:
         The full path of the output file as a string.
     """
     available_partition_files = get_downloaded_partitions(path_to_output_file.parent)
-    with path_to_output_file.open("wb") as outfile:
-        for file_path in available_partition_files:
-            with file_path.open("rb") as file_source:
-                shutil.copyfileobj(file_source, outfile, length=(5 * 1024 * 1024))
-            file_path.unlink()
+    if len(available_partition_files) != 0:
+        with path_to_output_file.open("wb") as outfile:
+            for file_path in available_partition_files:
+                with file_path.open("rb") as file_source:
+                    shutil.copyfileobj(file_source, outfile, length=(5 * 1024 * 1024))
+                file_path.unlink()
     return path_to_output_file.as_posix()
 
 
@@ -322,35 +338,173 @@ def get_files_ready_for_integrity_test(
 ##########################################################################################
 
 
-def pre_concatenation_processing(download_manifest: ConcurrentDownloadManifest):
+def pre_concatenation_processing(
+    download_manifest: ConcurrentDownloadManifest,
+) -> ConcurrentDownloadManifest:
+    """Implements the pre-concatenation processing phase.
+
+    The pre-concatenation processing phase consist in checking the files that have been
+    downloaded as a whole (because smaller than the multi-part threshold) for data
+    integrity first. The files that failed the data integrity test are added to a new
+    ConcurrentDownloadManifest that is used to collect the data of all those files or
+    partitions that need to be downloaded again. Once the whole files are checked, the
+    program proceeds with checking if all the partitions for all the files have been
+    downloaded. If any file has a missing partition, the file specific information of the
+    file is added to the ConcurrentDownloadManifest and the specific missing partitions
+    are added as well to be downloaded once again.
+
+    Parameters
+    ----------
+    download_manifest: ConcurrentDownloadManifest
+        A ConcurrentDownloadManifest named-tuple containing the download manifest that
+        was originally used to download the files concurrently.
+
+    Returns
+    -------
+    ConcurrentDownloadManifest
+        A ConcurrentDownloadManifest named-tuple containing the download manifest with
+        the information necessary to download the whole files that were incorrectly
+        downloaded in first place and the files that are missing some partitions.
+    """
+    failed_non_partitioned_files = get_list_of_failed_downloads(
+        get_non_partitioned_files(download_manifest.files_reference_data),
+    )
     missing_partitions = get_all_missing_partitions(
-        whole_files_reference_data=download_manifest.whole_files_reference,
-        concurrent_download_manifest=download_manifest.concurrent_download_manifest,
+        whole_files_reference_data=download_manifest.files_reference_data,
+        partitions_to_download=download_manifest.partitions_to_download,
     )
     files_with_missing_partitions = get_files_with_missing_partitions(
-        whole_files_reference_data=download_manifest.whole_files_reference,
+        whole_files_reference_data=download_manifest.files_reference_data,
         missing_partitions=missing_partitions,
     )
+    files_to_retry_reference_data = list(
+        itertools.chain(failed_non_partitioned_files, files_with_missing_partitions),
+    )
     return ConcurrentDownloadManifest(
-        whole_files_reference=files_with_missing_partitions,
-        concurrent_download_manifest=missing_partitions,
+        files_reference_data=files_to_retry_reference_data,
+        whole_files_to_download=failed_non_partitioned_files,
+        partitions_to_download=missing_partitions,
     )
 
 
 def concatenation_processing(
     download_manifest: ConcurrentDownloadManifest,
     failed_downloads_manifest: ConcurrentDownloadManifest,
-):
+) -> List[DownloadDetails]:
+    """Implements the concatenation processing phase.
+
+    During the concatenation processing phase, first all the information of all the files
+    that are not missing any partition is collected, and then it is passed to the
+    concatenating function that joins all the partitions of a specific-file to
+    create a single compressed file.
+
+    Parameters
+    ----------
+    download_manifest: ConcurrentDownloadManifest
+        The download manifest containing all the information used originally to download
+        the files.
+    failed_downloads_manifest: ConcurrentDownloadManifest
+        The download manifest containing the information of the files downloaded as a
+        whole that failed the data integrity test, and of the files that were split in
+        multiple partitions but that, after the initial download, were found missing one
+        or more partitions.
+
+    Returns
+    -------
+    List[DownloadDetails]
+        A list containing the DownloadDetails named-tuples of all the files that went
+        through the concatenation process.
+    """
+    files_with_missing_partitions = [
+        file for file in failed_downloads_manifest.files_reference_data
+        if file.is_partitioned is True
+    ]
     files_ready_for_concatenation = get_files_ready_for_concatenation(
-        whole_files_reference_data=download_manifest.whole_files_reference,
-        files_with_missing_partitions=failed_downloads_manifest.whole_files_reference,
+        whole_files_reference_data=download_manifest.files_reference_data,
+        files_with_missing_partitions=files_with_missing_partitions,
     )
-    concatenate_each_file_partitions(files_ready_for_concatenation)
+    return concatenate_each_file_partitions(files_ready_for_concatenation)
 
 
 def update_failed_download_manifest(
-    failed_download_manifest: ConcurrentDownloadManifest,
-    failed_files: List[DownloadDetails],
-    failed_partitions: List[PartitionDownloadDetails],
+    failed_downloads_manifest: ConcurrentDownloadManifest,
+    initial_download_manifest: ConcurrentDownloadManifest,
+    integrity_test_failing_files: List[DownloadDetails],
 ):
-    pass
+    """Updates the download manifest with the information of the files failing the integrity test.
+
+    Parameters
+    ----------
+    failed_downloads_manifest: ConcurrentDownloadManifest
+        The download manifest containing the information of the files downloaded as a
+        whole that failed the data integrity test, and of the files that were split in
+        multiple partitions but that, after the initial download, were found missing one
+        or more partitions.
+    initial_download_manifest: ConcurrentDownloadManifest
+        The download manifest containing all the information used originally to download
+        the files.
+    integrity_test_failing_files: List[DownloadDetails]
+        A list of DownloadDetails named-tuples of all those files that after concatenation
+        failed the data integrity test (e.g. because one of the partitions was not
+        downloaded properly).
+
+    Returns
+    -------
+    ConcurrentDownloadManifest
+        The download manifest containing the information of all those files whose
+        download failed for different reasons (incomplete download, incomplete download
+        of partitions, missing partitions etc.).
+    """
+    files_reference_data_to_update = copy.deepcopy(failed_downloads_manifest.files_reference_data)
+    partition_to_download_to_update = copy.deepcopy(
+        failed_downloads_manifest.partitions_to_download,
+    )
+    updated_file_reference = list(
+        itertools.chain(
+            files_reference_data_to_update,
+            integrity_test_failing_files,
+        ),
+    )
+    failed_partitions = []
+    for failed_file in integrity_test_failing_files:
+        file_specific_partitions = get_partitions_download_details(
+            initial_download_manifest.partitions_to_download,
+            failed_file.file_name,
+        )
+        failed_partitions.extend(file_specific_partitions)
+    updated_partitions_to_download = list(itertools.chain(
+        partition_to_download_to_update,
+        list(itertools.chain(failed_partitions)),
+    ))
+    return ConcurrentDownloadManifest(
+        files_reference_data=updated_file_reference,
+        whole_files_to_download=failed_downloads_manifest.whole_files_to_download,
+        partitions_to_download=updated_partitions_to_download,
+    )
+
+
+def post_concurrent_download_processing(
+    download_manifest: ConcurrentDownloadManifest,
+) -> ConcurrentDownloadManifest:
+    """Implements the post concurrent download processing phase.
+
+    Parameters
+    ----------
+    download_manifest: ConcurrentDownloadManifest
+        The download manifest containing all the information used originally to download
+        the files.
+
+    Returns
+    -------
+    ConcurrentDownloadManifest
+        The download manifest containing the information of the files that need to be
+        downloaded once again.
+    """
+    initial_failed_downloads = pre_concatenation_processing(download_manifest)
+    concatenated_files = concatenation_processing(download_manifest, initial_failed_downloads)
+    integrity_test_failing_downloads = get_list_of_failed_downloads(concatenated_files)
+    return update_failed_download_manifest(
+        initial_failed_downloads,
+        download_manifest,
+        integrity_test_failing_downloads,
+    )
